requirements.txt
aniso8601==9.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.2
Flask-RESTful==0.3.9
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
pytz==2022.7.1
requests==2.28.2
six==1.16.0
urllib3==1.26.14
uWSGI==2.0.21
Werkzeug==2.2.2
zipp==3.11.0
